rename clusterstatus resource to clusterinfrarefstatus
