+++ service/controller/resource/clusterinfrarefstatus/delete.go
@@ -1,4 +1,4 @@
-package clusterstatus
+package clusterinfrarefstatus
 
 import (
 	"context"
